+++ packages/web-components/src/components/global-header/components/global-header/src/components/HybridIpaasHeader/HybridIpaasHeader.ts
@@ -74,6 +74,7 @@ export class HybridIpaasHeader extends LitElement {
   @property({ type: Boolean }) addCookiePreferences = false;
   @property({ type: Boolean }) solisSessionManagerEnabled = false; // toggle to enable/disable the Solis session manager
   @property({ type: Number }) solisSessionRefreshInterval = 25; // might not need Solis token refresh interval to be configurable
+  @property({ type: Number }) solisIdleTimeoutInterval = 28; // might not need Solis idle timeout interval to be configurable
 
   @state()
   headerOptions: HeaderProps = {
@@ -108,6 +109,7 @@ export class HybridIpaasHeader extends LitElement {
     super.disconnectedCallback();
     if (this.sessionManager) {
       this.sessionManager.stopRefreshSchedule();
+      this.sessionManager.unregisterActivityListeners();
       this.sessionManager = null;
     }
     document.removeEventListener(CUSTOM_EVENT_NAME, this._customEventListener);
@@ -193,9 +195,11 @@ export class HybridIpaasHeader extends LitElement {
     if (!this.sessionManager) {
       this.sessionManager = new solisSessionManager({
         tokenRefreshInterval: this.solisSessionRefreshInterval,
+        idleTimeoutInterval: this.solisIdleTimeoutInterval,
         basePath: this.basePath,
       });
       this.sessionManager.startRefreshSchedule();
+      this.sessionManager.registerActivityListeners();
     }
   }