+++ packages/web-components/src/components/global-header/components/global-header/src/types/Header.types.ts
@@ -393,5 +393,6 @@ declare global {
 
 export interface solisSessionManagerConfig {
   tokenRefreshInterval?: number;
+  idleTimeoutInterval?: number;
   basePath?: string;
 }