+++ packages/web-components/src/components/wysiwyg/__tests__/__snapshots__/wysiwyg.test.snap.js
@@ -1,0 +1,18 @@
+/**
+ * Copyright IBM Corp. 2026
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+/* @web/test-runner snapshot v1 */
+export const snapshots = {};
+
+snapshots['clabs-wysiwyg should render with minimum attributes'] =
+  `<clabs-wysiwyg
+  content=""
+  toolbar-size="md"
+>
+</clabs-wysiwyg>
+`;
+/* end snapshot clabs-wysiwyg should render with minimum attributes */